+++ react-native/src/blueDarkAltTheme.ts
@@ -9,13 +9,12 @@ This code is licensed under the BSD-3 license found in the LICENSE file in the r
 import { blue, black, lightBlue } from '@pxblue/colors';
 import { blueDarkTheme } from './blueDarkTheme';
 
-export const blueDarkSupplementTheme: typeof blueDarkTheme = {
+export const blueDarkAltTheme: typeof blueDarkTheme = {
     ...blueDarkTheme,
     colors: {
         ...blueDarkTheme.colors,
         primary: blue[500],
         accent: lightBlue[500],
         background: black[900],
-        notification: lightBlue[500],
     },
 };